crud/forms.py
# users/forms.py
from django import forms
from crispy_forms.helper import FormHelper
from crispy_forms.layout import Layout, Submit

from crud.models import UserProfile
# from localflavor.generic.forms import IBANFormField


class UserProfileForm(forms.ModelForm):
    class Meta:
        model = UserProfile
        fields = '__all__'
