requirements.txt
Django>=2.0,<2.1
psycopg2-binary==2.7.5
django-extensions==2.1.2
Werkzeug==0.14.1
django-allauth==0.37.1
python3-openid==3.1.0
requests==2.19.1
oauth2==1.9.0.post1
django-localflavor==2.1
django-crispy-forms==1.7.2
